fastApiProject/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.25
pydantic==2.6.4
pytest==8.2.0
httpx==0.27.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
